test: reorganization of the 'ep' unit test
