+++ tests/unit/CMakeLists.txt
@@ -5,7 +5,7 @@
 
 add_subdirectory(conn)
 add_subdirectory(conn_req)
-add_subdirectory(ep-test)
+add_subdirectory(ep)
 add_subdirectory(error)
 add_subdirectory(flush-test)
 add_subdirectory(info-test)
